feat: -- 新增 SqlToy.

+++ 2.x/spring-boot-sqltoy/src/main/java/com/lmaye/sqltoy/entity/SysUser.java
@@ -7,6 +7,7 @@ import org.sagacity.sqltoy.config.annotation.Entity;
 import org.sagacity.sqltoy.config.annotation.Id;
 
 import java.io.Serializable;
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 /**
@@ -25,13 +26,13 @@ public class SysUser implements Serializable {
      */
     @Id(strategy = "generator", generator = "org.sagacity.sqltoy.plugins.id.impl.DefaultIdGenerator")
     @Column(name = "id", comment = "主键ID", length = 19L, type = java.sql.Types.BIGINT, nullable = false)
-    private Long id;
+    private BigInteger id;
 
     /**
      * 部门ID
      */
     @Column(name = "dept_id", comment = "部门ID", length = 19L, type = java.sql.Types.BIGINT, nullable = true)
-    private Long deptId;
+    private BigInteger deptId;
 
     /**
      * 客户端ID
